# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  analogdb.CatalogPost:
    properties:
      id:
        example: 1
        type: integer
      images:
        items:
          $ref: '#/definitions/analogdb.Image'
        type: array
      score:
        example: 150
        type: integer
      title:
        example: Sunset over the lake
        type: string
    type: object
...
  analogdb.CreateFilm:
    properties:
      color_type:
        example: color
        type: string
      description:
        example: Kodak Portra 400 is professional color negative film with fine grain
        type: string
      make:
        example: kodak
        type: string
      speed:
        example: 400
        type: integer
      type:
        example: portra 400
        type: string
    type: object
...
  analogdb.Error:
    properties:
      code:
        example: not_found
        type: string
      message:
        example: not found
        type: string
    type: object
  analogdb.Film:
    properties:
      color_type:
        example: color
        type: string
      created:
        example: "2025-07-11T12:00:00Z"
        type: string
      description:
        example: Kodak Portra 400 is professional color negative film with fine grain
        type: string
      id:
        example: 1
        type: integer
      make:
        example: kodak
        type: string
      post_count:
        example: 25
        type: integer
      slug:
        example: kodak-portra-400
        type: string
      speed:
        example: 400
        type: integer
      top_posts:
        items:
          $ref: '#/definitions/analogdb.CatalogPost'
        type: array
      type:
        example: portra 400
        type: string
      updated:
        example: "2025-07-11T12:00:00Z"
        type: string
    type: object
  analogdb.Image:
    properties:
      height:
        example: 7600
        type: integer
      resolution:
        example: high
        type: string
      url:
        example: https://cloudfront.net/1.jpeg
        type: string
      width:
        example: 4800
        type: integer
    type: object
...
  server.CaptionsMissingResponse:
    properties:
      ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    type: object
...
  server.CreateFilmResponse:
    properties:
      film:
        $ref: '#/definitions/analogdb.CreateFilm'
      message:
        example: Success, film created
        type: string
    type: object
...
  server.DeleteResponse:
    properties:
      message:
        example: Success, post deleted
        type: string
    type: object
...
  server.FilmsResponse:
    properties:
      films:
        items:
          $ref: '#/definitions/analogdb.Film'
        type: array
    type: object
...
  server.VectorsMissingResponse:
    properties:
      extra:
        example: 0
        type: integer
      ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    type: object
...
host: api.analogdb.com
info:
  contact: {}
  description: API for analogdb film database
  title: AnalogDB API
  version: "1.0"
paths:
...
  /film:
    post:
      consumes:
      - application/json
      description: Create a new film entry (requires authentication)
      parameters:
      - description: Film data to create
        in: body
        name: film
        required: true
        schema:
          $ref: '#/definitions/analogdb.CreateFilm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.CreateFilmResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/analogdb.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: Create a new film
      tags:
      - film
    put:
      consumes:
      - application/json
      description: Create a new film entry (requires authentication)
      parameters:
      - description: Film data to create
        in: body
        name: film
        required: true
        schema:
          $ref: '#/definitions/analogdb.CreateFilm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.CreateFilmResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/analogdb.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: Create a new film
      tags:
      - film
  /film/{id}:
    delete:
      description: Delete a catalog film by ID (requires authentication). Refused
        while any post still uses it.
      parameters:
      - description: Film ID to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.DeleteResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/analogdb.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/analogdb.Error'
        "409":
          description: Still used by posts
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: Delete a film
      tags:
      - film
  /films:
    get:
      consumes:
      - application/json
      description: Retrieve films with optional query parameters for filtering and
        sorting
      parameters:
      - description: Sort order
        enum:
        - alphabetical
        - counts
        in: query
        name: sort
        type: string
      - description: Number of results to return
        in: query
        name: page_size
        type: integer
      - description: Filter by film make
        in: query
        name: make
        type: string
      - description: Filter by film type
        in: query
        name: type
        type: string
      - description: Filter by film speed
        in: query
        name: speed
        type: integer
      - description: Filter by color type
        in: query
        name: colortype
        type: string
      - description: Filter by specific film ID
        in: query
        name: id
        type: integer
      - description: Include count data
        in: query
        name: include_counts
        type: boolean
      - description: Exclude zero counts
        in: query
        name: exclude_zero_counts
        type: boolean
      - description: Only return entries with at least this many posts, implies include_counts
        in: query
        name: min_count
        type: integer
      - description: Attach this many top scoring posts to each entry (1 to 10)
        in: query
        name: top_posts
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.FilmsResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      summary: Get films with optional filtering
      tags:
      - films
...
  /scrape/captions/missing:
    get:
      description: List ids of posts with no caption, or with a caption of another
        version when version is set (requires authentication)
      parameters:
      - description: Caption version the posts should have
        in: query
        name: version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CaptionsMissingResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: List posts missing a caption
      tags:
      - scrape
  /scrape/vectors/missing:
    get:
      description: List ids of posts with no object in the vector database, oldest
        first, and how many objects belong to deleted posts (requires authentication)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.VectorsMissingResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
        "503":
          description: Vector database unavailable
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: List posts missing a vector
      tags:
      - scrape
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"