# backend/server/films.go
package server

import (
	"net/http"

	"github.com/evanofslack/analogdb"
	"github.com/go-chi/chi/v5"
)

type FilmsResponse struct {
	Films []analogdb.Film `json:"films"`
}

type CreateFilmResponse struct {
	Message string              `json:"message" example:"Success, film created"`
	Film    analogdb.CreateFilm `json:"film"`
}

// default to sorting alphabetical
var defaultFilmsSort = analogdb.FilmSortAlphabetical

// max limit of films or cameras returned
var maxListLimit = 1000

// max top posts attached to each film or camera
var maxTopPosts = 10

const (
	filmsPath = "/films"
	filmPath  = "/film"
)

func (s *Server) mountFilmHandlers(r chi.Router) {
	r.Route(filmsPath, func(r chi.Router) {
		r.Get("/", s.getFilms)
	})
	r.Route(filmPath, func(r chi.Router) {
		r.With(s.require(roleScraper)).Put("/", s.createFilm)
		r.With(s.require(roleScraper)).Post("/", s.createFilm)
		r.With(s.require(roleScraper)).Delete("/{id}", s.deleteFilm)
	})
}

// @Summary Get films with optional filtering
// @Description Retrieve films with optional query parameters for filtering and sorting
// @Tags films
// @Accept json
// @Produce json
// @Param sort query string false "Sort order" Enums(alphabetical, counts)
// @Param page_size query int false "Number of results to return"
// @Param make query string false "Filter by film make"
// @Param type query string false "Filter by film type"
// @Param speed query int false "Filter by film speed"
// @Param colortype query string false "Filter by color type"
// @Param id query int false "Filter by specific film ID"
// @Param include_counts query bool false "Include count data"
// @Param exclude_zero_counts query bool false "Exclude zero counts"
// @Param min_count query int false "Only return entries with at least this many posts, implies include_counts"
// @Param top_posts query int false "Attach this many top scoring posts to each entry (1 to 10)"
// @Success 200 {object} FilmsResponse
// @Failure 400 {object} analogdb.Error "Invalid query parameters"
// @Failure 500 {object} analogdb.Error "Internal server error"
// @Router /films [get]
func (s *Server) getFilms(w http.ResponseWriter, r *http.Request) {
	filter, err := parseToFilmFilter(r)
	if err != nil {
		s.writeError(w, r, err)
		return
	}
	resp, err := s.makeFilmResponse(r, filter)
	if err != nil {
		s.writeError(w, r, err)
		return
	}
	err = encodeResponse(w, r, http.StatusOK, resp)
	if err != nil {
		s.writeError(w, r, err)
	}
}

func (s *Server) makeFilmResponse(r *http.Request, filter *analogdb.FilmFilter) (FilmsResponse, error) {
	films, err := s.FilmService.FindFilms(r.Context(), filter)
	resp := FilmsResponse{}
	if err != nil {
		return resp, err
	}
	for _, f := range films {
		film := *f
		film.Slug = analogdb.Slug(film.Make, film.Type)
		resp.Films = append(resp.Films, film)
	}
	return resp, nil
}

// @Summary Create a new film
// @Description Create a new film entry (requires authentication)
// @Tags film
// @Accept json
// @Produce json
// @Param film body analogdb.CreateFilm true "Film data to create"
// @Success 201 {object} CreateFilmResponse
// @Failure 400 {object} analogdb.Error "Invalid request body"
// @Failure 401 {object} analogdb.Error "Unauthorized"
// @Failure 403 {object} analogdb.Error "Forbidden"
// @Failure 422 {object} analogdb.Error "Unprocessable entity"
// @Failure 500 {object} analogdb.Error "Internal server error"
// @Security BasicAuth
// @Router /film [put]
// @Router /film [post]
func (s *Server) createFilm(w http.ResponseWriter, r *http.Request) {
	var createFilm analogdb.CreateFilm
	if err := s.decodeBody(w, r, &createFilm, "parse film from request body"); err != nil {
		s.writeError(w, r, err)
		return
	}

	created, err := s.FilmService.CreateFilm(r.Context(), &createFilm)
	if err != nil || created == nil {
		s.writeError(w, r, err)
		return
	}

	createdResponse := CreateFilmResponse{
		Message: "Success, film created",
		Film:    *created,
	}
	if err := encodeResponse(w, r, http.StatusCreated, createdResponse); err != nil {
		s.writeError(w, r, err)
	}
}

// @Summary Delete a film
// @Description Delete a catalog film by ID (requires authentication). Refused while any post still uses it.
// @Tags film
// @Produce json
// @Param id path int true "Film ID to delete"
// @Success 200 {object} DeleteResponse
// @Failure 400 {object} analogdb.Error "Invalid ID"
// @Failure 401 {object} analogdb.Error "Unauthorized"
// @Failure 403 {object} analogdb.Error "Forbidden"
// @Failure 404 {object} analogdb.Error "Not found"
// @Failure 409 {object} analogdb.Error "Still used by posts"
// @Failure 500 {object} analogdb.Error "Internal server error"
// @Security BasicAuth
// @Router /film/{id} [delete]
func (s *Server) deleteFilm(w http.ResponseWriter, r *http.Request) {
	id, err := stringToInt(chi.URLParam(r, "id"))
	if err != nil {
		s.writeError(w, r, err)
		return
	}
	if err := s.FilmService.DeleteFilm(r.Context(), id); err != nil {
		s.writeError(w, r, err)
		return
	}
	if err := encodeResponse(w, r, http.StatusOK, DeleteResponse{Message: "Success, film deleted"}); err != nil {
		s.writeError(w, r, err)
	}
}

// parse URL for query parameters and convert to FilmFilter
func parseToFilmFilter(r *http.Request) (*analogdb.FilmFilter, error) {
	filter := analogdb.NewFilmFilter(nil, &defaultFilmsSort, nil, nil, nil, nil, nil, nil, nil)

	values := r.URL.Query()

	if sort := values.Get("sort"); sort != "" {
		if sort == "alphabetical" || sort == "counts" {
			switch sort {
			case "alphabetical":
				alpha := analogdb.FilmSortAlphabetical
				filter.Sort = &alpha
			case "counts":
				counts := analogdb.FilmSortCounts
				filter.Sort = &counts
			}
		} else {
			return nil, badRequest("invalid sort parameter %s, valid options are 'alphabetical', or 'counts'", sort)
		}
	}

	if limit := values.Get("page_size"); limit != "" {
		if intLimit, err := stringToInt(limit); err != nil {
			return nil, err
		} else {
			intLimit = clampLimit(intLimit, 1, 1, maxListLimit)
			filter.Limit = &intLimit
		}
	}

	if make := values.Get("make"); make != "" {
		filter.Make = &make
	}

	if ty := values.Get("type"); ty != "" {
		filter.Type = &ty
	}

	if speed := values.Get("speed"); speed != "" {
		if intSpeed, err := stringToInt(speed); err != nil {
			return nil, err
		} else {
			filter.Speed = &intSpeed
		}
	}

	if color := values.Get("colortype"); color != "" {
		filter.ColorType = &color
	}

	if id := values.Get("id"); id != "" {
		if identify, err := stringToInt(id); err != nil {
			return nil, err
		} else {
			filter.IDs = &[]int{identify}
		}
	}

	if includeCounts := values.Get("include_counts"); includeCounts != "" {
		if val, err := stringToBool(includeCounts); err != nil {
			return nil, err
		} else {
			filter.IncludeCounts = &val
		}
	}

	if excludeZero := values.Get("exclude_zero_counts"); excludeZero != "" {
		if val, err := stringToBool(excludeZero); err != nil {
			return nil, err
		} else {
			filter.ExcludeZeroCounts = &val
		}
	}

	if minCount := values.Get("min_count"); minCount != "" {
		if val, err := stringToInt(minCount); err != nil {
			return nil, err
		} else if val < 1 {
			return nil, badRequest("invalid min_count parameter %d, must be at least 1", val)
		} else {
			filter.MinCount = &val
		}
	}

	if topPosts := values.Get("top_posts"); topPosts != "" {
		if val, err := stringToInt(topPosts); err != nil {
			return nil, err
		} else {
			val = clampLimit(val, 1, 1, maxTopPosts)
			filter.TopPosts = &val
		}
	}

	return filter, nil
}